+++ src/html/index.html
@@ -6,6 +6,7 @@
 	@@include( 'template-parts/sections/home/main-video.html', { data: 'https://player.vimeo.com/video/793672371?h=2f199533b5', 'src': 'img/home/video.png' } )
 	@@include( 'template-parts/sections/home/slides/slides.html' )
 	@@include( 'template-parts/sections/home/video-slides/video-slides.html' )
+	@@include( 'template-parts/sections/home/about.html' )
 </main>
 
 @@include( 'template-parts/footer/footer.html' )
