+++ src/html/template-parts/sections/home/about.html
@@ -1,0 +1,42 @@
+<section class="about">
+	<div class="container">
+		<div class="about-items">
+			<div class="about-info">
+				<h2>
+					Нальгиев Магомед Макшарипович
+				</h2>
+				<p class="about-subtitle">
+					Пластический хирург клиники эстетической медицины «Время красоты»
+				</p>
+				<div class="about-doctor">
+					<p class="about-doctor-item">
+						Опыт работы - 8 лет
+					</p>
+					<p class="about-doctor-item">
+						Все виды эстетических операций лица и тела
+					</p>
+					<h3>
+						Образование
+					</h3>
+					<ul class="about-list">
+						<li class="about-item">
+							<span>1</span> Ингушский государственный университет, медицинский факультет, квалификация «врач по специальности лечебное дело»
+						</li>
+						<li class="about-item">
+							<span>2</span> Интернатура по специальности «Общая хирургия» на базе Ингушской Республиканской клинической больницы
+						</li>
+						<li class="about-item">
+							<span>3</span> Клиническая ординатура РНИМУ им. Н. И. Пирогова по специальности «Пластическая хирургия»
+						</li>
+					</ul>
+					<a href="#" class="button white">
+						Задать вопрос доктору
+					</a>
+				</div>
+			</div>
+			<div class="about-img">
+				<img src="img/home/about/doc.png" width="609" height="864" alt="">
+			</div>
+		</div>
+	</div>
+</section>
