+++ src/js/main.js
@@ -6,6 +6,5 @@ import './header'
 import './fancyBox'
 import './lazy-frame'
 import './inputs'
-import './slides/slider'
 import './slides/slides'
