--- src/js/slides/slider.js
@@ -1,25 +1,0 @@
-
-// import { initSimpleSwiper } from "../common/global"
-
-// document.addEventListener( 'DOMContentLoaded', () => {
-// 	'use strict'
-
-// 	initSimpleSwiper( '.main-swiper', '.main-pagination', '.slider-wrapper',
-// 		{	
-// 			breakpoints: {
-// 				320: {
-// 					slidesPerView: 1,
-// 				},
-// 				576: {
-// 					slidesPerView: 2,
-// 				}
-// 			},
-// 		}
-// 	)
-
-// 	initSimpleSwiper( '.swiper.video-swiper', '.video-pagination', '.video-slides-wrapper', 
-// 		{	
-// 			slidesPerView: 1,
-// 		}
-// 	)
-// } )
