+++ src/scss/main.scss
@@ -24,3 +24,4 @@
 @import 'sections/frames';
 @import 'sections/home/slider';
 @import 'sections/home/video-slides';
+@import 'sections/home/about';
