+++ src/scss/sections/home/video-slides.scss
@@ -36,10 +36,6 @@
 		height: 614px;
 	}
 
-	&-left {
-	
-	}
-
 	&-right {
 		display: flex;
 		flex-direction: column;
